Add experimental -fnew-inlining-behavior (#4972)
This flag is _extremely_ experimental. It is the first step in an effort
to clean up and adjust our optimization pipeline ordering. This first
step enables compiling shaders using a more traditional LLVM pass
ordering and the normal inlining behavior. For library shaders it also
skips adding the `alwaysinline` attribute to all function calls, so we
end up only inlining profitable calls based on the inliner heuristics.

There is much more work to do to get this flag to generate valid
shaders, but adding the flag so that we can start working on it is the
first step.