Support dumping implicit top-level decls (#5007)
This change reworks how we were filtering top-level implicit decls from
the `-ast-dump` output. The change has a couple of key differences, but
the behavior of the `-ast-dump` flag remains unchanged.

The differences in this change are:
* By default Decl::dump _will_ dump implicit top-level decls which
allows attached debuggers to dump top-level decls.
* Top-level implicit decls are determined by the decl's `Implicit` bit
rather than an invalid source location.
* Disabling dumping implicit decls is now a language option, so it can
be toggled.
* A new `-ast-dump-implicit` flag is added to dump the AST including
implicit top-level decls.

The attached test case verifies that the `-ast-dump` flag does not dump
implicit top-level decls, and it is intended to be strict enough that
if we add new top-level decls in the future and fail to mark them as
`implicit` the test will fail.

The attached test also trys to fill some gaps in testing around the
lazy decl initialization by verifying that incomplete implicit decls
are generated, and that a used decl gets completed as expected.