Don't add template record decl to declcontext (#5016)
CXXRecordDecls that are part of ClassTemplateDecls should not be added
to the DeclContext that the ClassTemplateDecl gets added to.

I'm a little surprised this didn't cause problems somewhere, but it
does result in the decl being dumped twice in the ast dumping. I think
it should have also resulted in conflicting decls in name lookup, but
there may be some other change somewhere that worked around that, or it
might get ignored since the decl is invalid by itself.

Either way, we can skip adding CXXRecordDecls for template types to the
DeclContext and everything works fine.

Function and Method declarations similarly don't need to be added
if they are children of FunctionTemplateDecl. This did cause duplicate
and odd diagnosticformulations