Introduce HLSLCXXOverload attribute (#5018)
This change should have no functional impact on compilation. The goal is
to remove cases where we rely on the subscript operator's name for
special resolution rules and instead key the behavior off an attribute
on the method declaration.

The added test cases are not exhaustive but try to cover at least one
object type exercising each code path modified in this change.

Also added a test case for HLSL 2021 user defined array operator. This
change partially addresses #4340. The remaining parts of that issue
will be addressed in a future HLSL language revision.