Always create temp for output signature
This is for case where output signature is used as argument for functions call.
A tmp is needed because the signature will be removed in dxil.

Doing it in LegalizeDxilInputOutputs is to reused existing feature which create the tmp.
Only add the temp for signature when there's real function call.