Refine maybe_unused
This refines the maybe_unused annotation, adopting C++ naming and more
C++-like behavior.

When applied to an `out` parameter `maybe_unused` results in skipping
uninitiailized value analysis on return and function block exits only
if the value is always uninitialized.

If the value is maybe uninitialized or if the value is used explicitly
the warnings are still emitted.