Break test dependencies on mangling (Part 2)
This is the second part of my changes to break test dependencies on
symbol mangling.

This change rewrites a handful of the tests in ValidationTest.cpp that
rely on rewriting IR from compiled shaders to be invalid. The new
versions of the tests use simplified IR as the source instead of HLSL.
Each test has the HLSL source it was derived from captured in code
comments.