Add diagnostics for uninitialized `out` parameters (#5047)
This change adds diagnostic support for uninitialized `out` parameters.
With this change, `out` parameters are treated as uninitiailzed values
until they are initialized and use of the values produdces a diagnostic.

This change is a slightly hacky extension of Clang's uninitiailzed
values analysis which is achieved by adding ParamVarDecls for `out`
parameters to the list of variables to analyize. It also adds dummy
uses of the variables on `return` statements.

It also introduces the maybe_unused attribute, adopting C++ naming and
C++-like behavior.

When applied to an `out` parameter `maybe_unused` results in skipping
uninitialized value analysis on return and function block exits only
if the value is always uninitialized.

If the value is maybe uninitialized or if the value is used explicitly
the warnings are still emitted.