Only track scalar, vector and struct `out` params
I should have realized this before, but the UninitializedValues
analysis doesn't correctly handle arrays or pointer objects. Because of
HLSL's value semantics we do get the expected results for scalar
parameters, and we get correct (but incomplete) results for vector and
struct parameters.

This limits the warning to only apply to scalar, vector, and struct
parameter types where the analysis is correct, and disables it for
arrays, pointers, and other types where it will not be correct.

This change also disables the "suggest initialization" fix-it for `out`
parameters. Default initiailiztion for `out` parameters doesn't work,
and would be confusing.