Merged PR 5226: [DXIL] Add validation for CompilerVersion Dxil Part
After adding the new part to the library dxil containers, there was a lack of validation that the new part was constructed correctly. This PR aims to address this problem by adding a function to dxilvalidation.cpp to validate the new CompilerVersion part. With the validation in dxilvalidation.cpp, the container tests can now assume / rely on the validator that the part is well-formed, when making value checks on the parts.