Only track scalar `out` params (#5078)
The UninitializedValues analysis doesn't correctly handle arrays or pointer objects, so we should not enable out parameter analysis for those types. Additionally, because the HLSL ASTs incorrectly represent structure out parameters as non-reference types, the analysis will generate incorrect diagnostics for uses of structs in call expressions. There isn't really an easy solution for that. This change takes a conservative approach to addressing those issues by disabling the analysis for values that we can't correctly identify.

This change also disables the "suggest initialization" fix-it for out parameters. Default initialization for out parameters doesn't work, and would be confusing.