spirv: fix GL semantic in inherited classes
When a semantic was in an inherited class, it was missed and caused
issues later down the pipe. This was because we assumed semantic fields
would only be present at the top-level struct, ignoring inheritance.

This commit also changes the parameter type from `DeclaratorDecl` to
`NamedDecl` as it only requires methods up to `NamedDecl`. Loosening
this requirement allowed me to pass a CXXRecordDecl.

Fixes #5138

Signed-off-by: Nathan Gauër <brioche@google.com>