ci: add GCP pipeline files to replace AppVeyor
Those files are the required to drive the new GCP pipeline aimed to
replace AppVeyor. As-is, those file have no effect on the repo.
What's needed to add the CI is to setup some repository permissions,
and only then CI will be able to run.

This CI will build x86_64 versions of DXC on both Linux and Windows, run
tests, and allow retrieving build artifacts (same as Appveyor).

Signed-off-by: Nathan Gauër <brioche@google.com>