Enable distribution builds through Visual Studio
This gets the `distribution` target working for DXC through visual
studio builds, and with the the `hctbuild` tooling.

Since DXC has a very limited number of targets that it relies on for
packaging this works in IDE builds (unlike LLVM where the number of
targets is insane).

This also updates `hctbuild` to have a flag for specifying the install
path, and sets the default install path to be `%HLSL_BLD_DIR%\install`,
nesting local install bits under the build directory.