test: add static assert for the git count define
The `target_version_vertex.hlsl` test looks at different defines
values, and generate the good or the bad instruction depending
on those defines.

This means an HLSL test will fail, because you are building DXC from
a shallow clone.
This commit adds a static assert if you build tests on a shallow clone
to let you fail earlier, with a more user-friendly explanation.