Set the layout rule in processTemplatedLoadFromBuffer (#5192)
* Set the layout rule in processTemplatedLoadFromBuffer

They layout rules for value coming from processTemplatedLoadFromBuffer
are inconsistent, and this is causing problems. This PR changes the
function to return an instruction whose layout rule is Void.

This should remove the inconsistenty.

Fixes #5123

* Add test

The cg.return.memory-layout.hlsl had to be updated because the Byteaddress load no longer has a different layout rule.

* Remove `-Vd` from test