PSV0: Fix GS multi stream out ViewID dependency serialization
GetViewIDOutputMask and GetInputToOutputTable take an output index to support multiple stream outputs in geometry shaders.
Previously, these functions would map the same local pointer to a location for serializing or deserializing the dependency bit vectors.
This causes data to overlap and size to be computed based only on the last used stream.

This change fixes this by using an array of pointers for each of these.
Then, a new test verifies output for two streams, which fails without this fix.