spirv: fix forward reference order for composite
The blanket rule for NonSemantic.Shader.DebugInfo.100 is "no forward
references". But there is a type-specific rule: DebugTypeComposite
members can be forward references:

```
Members must be the <id>s of DebugTypeMember, DebugFunction, or
DebugTypeInheritance. This could be a forward reference.
```

This is mandatory to correctly support HLSL class methods:
 - the composite has a function as member.
 - the function as the composite as scope.
This cycle makes the forward reference mandatory, hence the exception
in the spec.

The removed code wrongly implemented the general rule, ignoring this
exception.

This change doesn't just solves the cycle issue, but might add forward
references that were avoided (since we never visit members). But as far
as the spec goes, this is a valid behavior. So I believe there is no
need to add any additional logic to avoid forward references.

This PR depends on the validator to correctly validate those forward
references (KhronosGroup/SPIRV-Tools#5230).

Fixes: #4911.

Signed-off-by: Nathan Gauër <brioche@google.com>