Enhance `globallycoherent` mismatch diagnostics (#5121)
This change moves the `globallycoherent` mismatch diagnostics out of
code generation and into semantic analysis. Additionally the diagnostic
is enhanced to provide more information in the text about the mismatch,
and to fire in source locations more directly attributable to the
location of the mismatch.

This change also allows the `globallycoherent` annotation to apply to
functions, where it applies to the return type of the function, and
mismatches can be diagnosed on assignment of the return value and on
the return statements.

Fixes #4537.