Add a test dxil lib for testing
This creates a dxil library that doesn't sign anything, but does perform
the validation very much as the internal validator does. This is used to
test that the compiler and validator objects can properly locate, load,
and make use of the external dxil library.

The DxilLib test verifies both that loading the compiler with the
external dxil library available will load that library by default,
resulting in compiles without the usual warning when it is not found and
also that explicitly loading the library can be done and the correct API
calls can be used.

Fixes #5394