Allow function template default parameters in HLSL (#5234)
Default parameters for function templates is a C++11 feature. Since HLSL
is defined as C++03 using default parameters for function templates
produces a language extension warning. There is really no need for the
warning since the template syntax is disallowed in HLSL 2018 and earlier
and we effectively always support this syntax in 2021.

This change suppresses the warning by treating HLSL as C++11 when
diagnosing default template arguments. This changes the warning to a
C++98 compatability warning, which is only enabled if explicitly
requested.

In the future we should consider making future HLSL language versions
based on a more modern C++ to avoid issues like this.

Fixes #5221