Fixed entry point omitted from the PDB when it's the same as input filename (#5236)
The args in the PDB and debug DXIL are based CodeGenOptions::HLSLArgs. This list was created by copying the raw unprocessed arg list and excluding any args equal to the input file name. When the entry point happens to be the same as the input filename, it would get excluded from this list and the PDB and debug module end up with corrupted arg list where -E has no value or wrong value.

Fixed by only excluding args that are InputClass.