Revert "[DxbcConverter] Fix corruption of ICB integer values (#4790)" (#5253)
This reverts commit 0a1f7a19f4324642bde0a92bf49ecd41011e60a7.

Use an array of 32-bit integer values instead of 32-bit float values for the "dx.icb" immediate constant buffer data.
This prevents the FPU flipping the 22nd bit of integer constants, when their bit pattern represents a float32 signalling NaN on x86 architecture.

This change is causing rendering issues on PCs with AMD GPU and is blocking an upcoming Windows release.

The target branch of this revert is a release branch based on release-1.7.2212 at commit 83f8c6c5 and it will be used for the upcoming Windows release. The dxilconv change stays in main, at least for now.