Fix debug info nondeterminism in SROAGlobalsAndAllocas from staticGVs set
staticGVs is used to track which globals are static and don't need new debug info added at a certain point.
Some global values are erased during iteration of the work list, including globals that could be in this set.
This could lead to stale pointers in the set, which could alias new global values from SROA, nondeterministically causing missing debug info.

This change removes globals from the set (if present) in each path where they are deleted.
It also skips adding globals to the set (and MergeGepUse) that have no users in the first place, since these will just be deleted and have to be removed from the set later.