Remove unneeded assert (#5256)
When processing the dot intrinsic, there is an assert that ensure the
return type is the same as the componentent type of the operands. This
is not true when the types were originally half. At that point, the
operands will have been promoted to float, but the return type will be
promoted later.

Since the assert is not universally true, I will remove it.

Fixes #5048