Added "partial-lifetime-markers" to only generate lifetime.start (#5267)
- Added `-opt-enable partial-lifetime-markers` to only generate `@llvm.lifetime.start` and not `@llvm.lifetime.end`.

`@llvm.lifetime.start` can sometimes be helpful for codegen, but inserting `@llvm.lifetime.end` in the right places sometimes requires modifying cfg in ways that could negatively impact codegen quality. This change adds a way to generate only `@llvm.lifetime.start` and not `@llvm.lifetime.end` for all the benefits it provides without the problems of inserting `@llvm.lifetime.end`.

The added `-opt-enable partial-lifetime-markers` is off by default, and only kicks in when lifetime-markers are enabled.

In cases where generating `@llvm.lifetime.end` is easy (around function calls for arguments), they are generated as normal.

-----------------------------------------------------

In certain cases, lifetime markers can help reduce register pressure by reducing lifetime of values. In this example:
```
  for (int i = 0; i < N; ++i) {
    S state;
    if (conds[i]) {
      state.x0 = input[i+0];
    }
    if (conds2[i]) {
      output[i+0] = state.x0;
    }
  }
```
`State state` is declared inside the loop, but the alloca is effectively "hoisted" to the entry block, and the compiler has no knowledge that no value of `state` flows between loop iterations. The codegen ends up looking like this:
```
  loop:
    %last_iteration_val = phi float [ %new_val, %if.end.b ], [ undef, <preheader> ] ; <--- Value from previous iteration.
    br i1 %same_cond, %if.then.a, %if.end.a
  
  if.then.a:
    %new_val.then = ...

  if.end.a:
    %new_val = phi float [ %last_iteration_val, %loop ], [ %new_val.then, %if.then.0 ]
    br i1 %same_cond, %if.then.b, %if.end.b

  if.then.b:
    store %new_val
    br %if.end.b

  if.end.b:
    br i1, <loop-exit>, %loop
```

`mem2reg` has special logic for handling `lifetime.start`, where it basically assumes it as a def for the alloca. With lifetime-markers turned on, the IR ends up looking like this:

```
  loop:
    // Notice the phi is gone.
    br i1, %if.then.a, %if.end.a
  
  if.then.a:
    %new_val.then = ...

  if.end.a:
    // The undef is now here!
    %new_val = phi float [ undef, %loop ], [ %new_val.then, %if.then.0 ]
    br i1 %same_cond, %if.then.b, %if.end.b

  if.then.b:
    store %new_val
    br %if.end.b

  if.end.b:
    br i1 %same_cond, <loop-exit>, %loop

```

When `state` is a big struct, and when the loop has many temporary values, the phis for previous loop iterations can cause very large register pressure.