Fix special case with vector swizzle (#5281)
When handling the swizzles in the SPIR-V backen, there is a special case when
the final result returns the original vector. It tries to avoid adding the
vector shuffle instruction in spir-v. However, that path calls `doExpr` instead
of `loadIfGLValue` like the other path. This can sometimes cause a load to be
omitted.

This is changed so that `loadIfGLValue` is called for both cases.

Fixes #5275