Add missing NodeDispatchGrid and NodeMaxDispatchGrid checks
Broadcasting launch nodes must have either the NodeDispatchGrid
or the NodeMaxDispatchGrid attribute, but not both. These checks
were missing. This change adds error diagnostics for these cases,
and adds verifier tests.

Updated a number of tests that failed with the introduction of
these checks, as they did not obey the newly enfored constraints.

Also extended the verifier test to check NodeTrackRWInputSharing,
and FinishCrossGroupSharing diagnostics to replace FileCheck tests.