Handle size 1 vectors in getTypeWithCustomBitwidth
Vectors of 1 element are handled as scalars in the spir-v backend
because the element count of [vectors in
spir-v](https://registry.khronos.org/SPIR-V/specs/unified1/SPIRV.html#OpTypeVector)
must be at least 2.

This case is not handled in getTypeWithCustomBitwidth and is causing an
assert. The fix is to add code to explicitly handle vectors of size one.

Fixes #5130