Add cast after RWByteAddressBuffer atomic compare exchange (#5295)
* Add cast after RWByteAddressBuffer atomic compare exchange

The type of the RWByteAddressBuffer is always uint in the SPIR-V
representation. This means that the OpAtomicCompareExchange instruction
must have parameters and a result of type uint. If the original code
used an in for the result, then there is a type mismatch when storing
the value. This is fixed by add a cast when appropriate.

Fixes #4741