Emit Offset decoration for PhysicalStorage structs
Depending on the struct layout, the `Offset` decoration
is not always required.
When this storage class is setup, the decoration needs to be added.

The tricky part if the mix of AST types and SPIRV types with bitfields:
 - in the AST, bitfields are represented as their own fields.
 - in SPIRV, bitfields are meld into 1 field when possible.

This means we need to lower the type to SPIR-V to compute the correct
offset.
This would be fine if we could always generate the Offset decoration,
but some types like image samplers are structs, and computing an
offset for them is not clean (maybe no legal?).

So I wanted to wait to know the storage type to determine if
the decoration needed to be emitted of not.