Update build artifact permissions
I found other users reporting a similar issue:

https://github.com/actions/deploy-pages/issues/188

The root cause seems to be that the deploy-pages action changed to stop
correcting permissions problems in the upload since it was slowing down
deployment. The recommended fix is to set the permissions before
uploading the artifact.

Let's see if this works.