Handle passing RWStructureBuffer array to function
getFinalACSBufferCounter currently returns a CounterIdAliasPair. This is
because all cases in the original design could return the entire
counter variable. When using arrays for RWStructuredBuffers, we will
need to be able to return something the represents a portion of the
array that is needed. The best method of doing that is to represent to return a
SpirvIinstruction that is the variable with possibly some OpAccessChains
that points to the part of the array that is needed.

This is then used to allow copying individual counter variables.

It does not handle copying the entire array of counter variables.