[Matrix type] Add AST matrix type.
An AST matrix type is added based on upstream matrix type.
This is prepare for support matrix orientation in template.


matrix<float, 2, 3> will generate a AST matrix type instead of a TemplateSpecialization.
ArraySubscriptExpr will be created for matrix subscript instead of a CallExpr.
Rewriter will keep const and default parameter for matrix.
Error message will be different for matrix is not a class in AST anymore.

The rest will keep the same as before.

Next PR will add orientation to the AST matrix type.
Many matrix related code will be useless after this change.
Will remove the useless code in a separate PR.