NFC: add null-check before deref.
When DXC emits an error, the rest of the expression is
still parsed until we go back to the root and show the error.

This means some functions could return a nullptr, which is not
accounted for in many places.
The subsequent change will add a case where is breaks, so this commit
adds a few additional null-checks.

Signed-off-by: Nathan Gauër <brioche@google.com>