Refactor type lowering for local RWStructuredBuffers
To handle a local RWStructuredBuffers, we treat the local varible like a
pointer to the global variable. This is needed to properly represent the
semantics of these local variables. The difficulty is that this spread
out in a number of places creating special cases are not always
intuiative. It is also causing problems for arrays of
RWStructuredBuffers.

This change will modify the type of the local variable generated by the
spirv emitter to be a pointer instead of waiting for type lowering to
handle it.