SPIRV: don't generate invalid debug instructions
When generating debug instructions for a function, each
one is linked to its scope.
In case of member functions, this scope is the class.

When declaring a class, all its member, including functions must
be declared.

This cycle requires a forward reference, which is allowed by the
debug instruction spec, but not by parents: NonSemantic & SPIR-V specs.
This is a spec issue we have to fix. In the meantime, we decided to
emit a warning, and generate slightly worse debug instructions.

Context: https://github.com/KhronosGroup/SPIRV-Registry/issues/203

Signed-off-by: Nathan Gauër <brioche@google.com>