Remove 'using namespace std' to fix __noop ambiguity with latest libc++ (#5406)
A recent change to libc++ added a 'struct __noop' to the 'std'
namespace. When building with clang-cl using libc++ on Windows, this now
introduces an ambiguity between this std::__noop struct, and the MSVC
__noop intrinsic, in translation units where 'using namespace std;' is
used. This commit replaces these with using clauses for the exact
symbols required in the non-test cpp files.