Fixed a crash when linker erases functions that still have users. (#5437)
When linking to library with a list of `-exports`, the linker erases
functions that are not exported assuming they are no longer used. This
assumption is not correct when the function is a constructor for a
global variable. This change fixes the crash by first checking whether
the function still has uses before erasing.

This change adds a regression test with a non-trivially removable global
variable constructor that would crash without this change.