Instantiate argument default parameters
This addresses a latent bug in Clang that was fixed _way_ after DXC was
forked. We encounter it in DXC because of the way that we inject AST
bits through the external sema source, but it likely wasn't exposable
through C++.

This change is mostly a back porting of the following upstream LLVM
commits:

llvm-project/73c6a2448f24
llvm-project/f721e0582b15
llvm-project/c601377b2376
llvm-project/4409a83c2935

The end result of this change is that we still lazily instantiate
default arguments, but we do force them to be instantiated during sema
when they are used.

Fixes #5145