[lit] fall back to taef for taef only test options (#5404)
It is not easy to pass through the quoted /p options thru python. And
these options are mainly for local debug filter, not for taef FileCheck
tests. Just fall back to taef for these options and treat them like taef
helper.

This is help to allow lit FileCheck shell test.