Fix attribute collision for HL intrinsics
HL Intrinsic functions share declarations with those that match group and function signature, regardless of the original intrinsic name.
This means that intrinsics with differing attributes can be collapsed into the same HL functions, leading to incorrect attributes for some HL intrinsics.

This fixes this issue by adding the attributes to the HL operation mangling, the same way this issue was fixed for the HLWaveSensitive attribute before.