[lit] [SPIRV} Copy legal-examples to lit (#5320)
Copy SPIRV tests under legal-examples and modify to work with lit.

This is first PR for moving SPIRV test to lit.
The reason to copy instead of moving the tests is to make sure not lost
coverage before all tests switch to lit. And update all tests in one PR
will be a PR with 1000 files.