[Verifier] Add verifier tests to catch MaxRecord errors earlier (#5484)
This PR is one of several more PR's that move diagnostic emission
earlier for specific errors, so that the verifier can catch these
errors. This PR focuses on the MaxRecord* attributes applied on
parameters. Specifically, there are 2 behaviors that this PR will allow
DXC to detect earlier:

1. Prevent more than one of {MaxRecord, MaxRecordsSharedWith} from being
applied to a single parameter
2. Prevent MaxRecordsSharedWith from referencing the same parameter name
that the attribute is being applied to.

Some tests have been merged or deleted to consolidate these verifier
tests. There is still some behavior that cannot be tested at this time
(verifying that MaxRecordsSharedWith doesn't reference an invalid
parameter name), which will be addressed in the future. It's likely that
this location is too early to properly validate this behavior, and the
tests for this behavior may need to be left at the same location.

This PR is one of a few that will help implement this task:
https://github.com/microsoft/DirectXShaderCompiler/issues/5368