[C++17] Merge Replace uses of std::iterator with explicit using (#5485)
Merged from
https://github.com/llvm/llvm-project/commit/0a92aff721f43406691e38a0965fd0c917121d09

This patch removes all uses of `std::iterator`, which was deprecated in
C++17. While this isn't currently an issue while compiling LLVM, it's
useful for those using LLVM as a library.

For some reason there're a few places that were seemingly able to use
`std` functions unqualified, which no longer works after this patch.
I've updated those places, but I'm not really sure why it worked in the
first place.