Fix debug info generation for `this` (#5534) (#5536)
My recent changes to correct AST inaccuracies around `this` objects
caused crashes in debug info generation for `this`. This resolves the
crashes by making the code behave correctly when `this` is a reference.

This has no change to debug info output. The added test case should
verify the same output we've always had, but is an important addition
since this was not tested in our pre-merge testing.

Fixes #5530

Cherry-pick of dc82ed29c9330127341f1419ac99936be369bd8f