Add `-HV 2021` to test that uses templates
A recent fix (#5534) modified this test by removing the `-HV 2021` flag.
That change was fine because flag is not needed now that 2021 is
the default language version.

However, we have a downstream consumer of this repo where we have not
yet migrated to 2021 as the default language version. Rather than
modifying the test in the downstream repo it was suggested that we add
the flag back in the upstream repo to minimize churn.